+++ db/migrate/20190326211254_remove_import_columns_from_designs.rb
@@ -1,0 +1,9 @@
+class RemoveImportColumnsFromDesigns < ActiveRecord::Migration[6.0]
+  def change
+    remove_column :designs, :import_started_at, :datetime
+    remove_column :designs, :import_ended_at, :datetime
+    remove_column :designs, :rows_imported, :integer, null: false, default: 0
+    remove_column :designs, :total_rows, :integer, null: false, default: 0
+    remove_column :designs, :csv_file, :string
+  end
+end